created models_v1 and models_v2 directories and moved files
