+++ examples/imagenet/models_v1/googlenetbn.py
