+++ examples/imagenet/models_v1/nin.py
