+++ examples/imagenet/models_v2/googlenet.py
