+++ examples/imagenet/models_v2/googlenetbn.py
