+++ examples/imagenet/models_v2/nin.py
